Remove mobile/main.ts: backend main is already mobile-only (#322)
When the backend's main.ts was originally written in a different context, it supported mobile, desktop, and web. Because of this, some of the functionality was placed in mobile/main.ts. Since it has been mobile-only for a long time, get rid of mobile/main.ts and put the functionality directly in main.ts.